project discussion comments as long.

+++ data/src/main/java/com/proptiger/data/mvc/ProjectController.java
@@ -93,7 +93,7 @@ public class ProjectController extends BaseController {
     @RequestMapping("data/v1/entity/project/{projectId}/discussions")
     @ResponseBody
     @DisableCaching
-    public ProAPIResponse getDiscussions(@RequestParam(required = false) Integer commentId, @PathVariable int projectId) {
+    public ProAPIResponse getDiscussions(@RequestParam(required = false) Long commentId, @PathVariable int projectId) {
         List<ProjectDiscussion> comments = projectService.getDiscussions(projectId, commentId);
         return new ProAPISuccessResponse(super.filterFields(comments, null));
     }
